change database structure

+++ src/main/java/codeenthusiast/TrainingCenterApp/exercise/strengthexercise/StrengthExercise.java
@@ -3,6 +3,7 @@ package codeenthusiast.TrainingCenterApp.exercise.strengthexercise;
 import codeenthusiast.TrainingCenterApp.constants.RepetitionUnit;
 import codeenthusiast.TrainingCenterApp.constants.WeightUnit;
 import codeenthusiast.TrainingCenterApp.exercise.Exercise;
+import codeenthusiast.TrainingCenterApp.exercise.strengthexercise.pace.Pace;
 import codeenthusiast.TrainingCenterApp.trainingsession.TrainingSession;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -22,12 +23,12 @@ public class StrengthExercise extends Exercise {
     @Enumerated(EnumType.STRING)
     private RepetitionUnit repetitionUnit;
 
-    private int reps;
+    private short reps;
 
     @Enumerated(EnumType.STRING)
     private WeightUnit weightUnit;
 
-    private int weight;
+    private double weight;
 
     @Embedded
     private StrengthExerciseDetails strengthExerciseDetails;
